Day 4 Completed Functionality

+++ src/components/librarian/ListOfAvailableBooks.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import UserContext from "../../utils/UserContext";
 import { getDataAvailable } from "../helper/apicalls";
-
-
+import { Link } from "react-router-dom";
+import {ToastContainer, toast } from "react-toastify";
 const ListOfAvailableBooks = () => {
   const context = useContext(UserContext);
 
@@ -28,33 +28,43 @@ const ListOfAvailableBooks = () => {
   const books = "books";
 
   const handleAddToCart = (item, id) => {
-    const updatedCartData = [...cartData, item];
-    setCartData(updatedCartData);
-    let email = context.user?.email;
-    console.log(updatedCartData);
-  
-    let cartObject = {
-      "bookname": updatedCartData[0].name,
-      "author": updatedCartData[0].author,
-      "category": updatedCartData[0].category,
-      "member-email": email
+    
+
+      const updatedCartData = [...cartData, item];
+      setCartData(updatedCartData);
+      let email = context.user?.email;
+      console.log(updatedCartData);
+    
+      let cartObject = {
+        "id": item.id,
+        "name": item.name,
+        "author": item.author,
+        "category": item.category,
+        "quantity": item.quantity
+      }
+      console.log(item);
+    
+      // Retrieve existing cart data from local storage, if any
+      const existingCartData = JSON.parse(localStorage.getItem("cartData")) || [];
+    
+      if(existingCartData.length < 3){
+      // Add the new cart object to the existing cart data array
+      const updatedCartDataArray = [...existingCartData, cartObject];
+    
+      // Store the updated cart data array into local storage as a stringified JSON object
+      localStorage.setItem("cartData", JSON.stringify(updatedCartDataArray));
+    
+      /* disable button */
+    
+      setButtonStates((prevStates) => ({
+        ...prevStates,
+        [id]: true,
+      }));
+    }else{
+    //  alert("You have reached the maximum book allowed for borrowing")
+      toast("Reached the limit")
     }
-  
-    // Retrieve existing cart data from local storage, if any
-    const existingCartData = JSON.parse(localStorage.getItem("cartData")) || [];
-  
-    // Add the new cart object to the existing cart data array
-    const updatedCartDataArray = [...existingCartData, cartObject];
-  
-    // Store the updated cart data array into local storage as a stringified JSON object
-    localStorage.setItem("cartData", JSON.stringify(updatedCartDataArray));
-  
-    /* disable button */
-  
-    setButtonStates((prevStates) => ({
-      ...prevStates,
-      [id]: true,
-    }));
+    
   };
   
 
@@ -77,6 +87,19 @@ const ListOfAvailableBooks = () => {
       {/* <Link to="/cart">Borrow</Link> */}
       <div className="d-flex flex-column min-vh-100">
         <div className="container">
+        <>
+        {context.user?.role === "librarian" && (
+            <nav className="nav nav-tabs mb-4">
+              <Link to="/listofavailable" className="nav-link">
+                Available Books
+              </Link>
+              <Link to="/listofdamaged" className="nav-link">
+                Damaged Books
+              </Link>
+            </nav>
+           )}
+            {/* <Outlet /> */}
+          </>
           <div className="my-4 text-center">
             <h1 className="bg-default">List Of Available Books</h1>
           </div>
@@ -88,7 +111,9 @@ const ListOfAvailableBooks = () => {
                 <th>Author</th>
                 <th>Category</th>
                 <th>Language</th>
-                <th>Action</th>
+                {context.user?.role === "member" && (
+                    <th>Action</th>
+                    )}
                 {/* <th>ISBN</th> */}
                 {/* <th colSpan={2}>Action</th> */}
               </tr>
@@ -101,7 +126,7 @@ const ListOfAvailableBooks = () => {
                     <td>{row.author}</td>
                     <td>{row.category}</td>
                     <td>{row.language}</td>
-
+                 
                     {/* <td>{row.isbn}</td> */}
 
                     {/* <div
